Dev/api setup (#10)
* Add SaveJsonLambdaStack class in lambda/app.py

This commit introduces the SaveJsonLambdaStack class in the lambda/app.py. The new class contains an AWS Lambda function for saving JSON to a table and an API Gateway with a POST method associated with the Lambda function. This enhancement allows the program to accept JSON data via an API and save it to a designated AWS database table.

* Update JSON request template in api_gateway_integration

This commit modifies the "application/json" request template in the api_gateway_integration of lambda/app.py. The new format no longer wraps the request input in a "payload" object, but directly converts the input as a JSON string. This change simplifies the request payload structure.

* Add Python script to save JSON to AWS table

The commit introduces a new Python file, 'save_json_to_table.py', which takes a JSON event and a table name, converts the event into a Pandas DataFrame, and then saves it to the specified AWS table. This addition will simplify the process of storing event data in AWS and improve the system's efficiency.

* Add API key and usage plan to API Gateway

The commit adds an API key requirement and a usage plan to the AWS API Gateway configuration in the app.py file. The usage plan includes throttling settings to manage request rates and a burst limit. This could boost security and also prevent misuse of the API by limiting the number of requests.

* Add deployment stage to usage plan

The commit adds a deployment stage to the usage plan in the API Gateway configuration within the app.py file. This extends control over the specific operational stages where the usage plan is implemented which can help manage throttling strategy and better handle API requests.

* Print event and partition data by timestamp

The commit includes a print statement to log the event in the lambda_handler method. Also, it modifies the way data is stored in the S3 bucket: instead of attaching a timestamp to each row, the data is now partitioned by the 'timestamp' column.

* Add automated lambda deployment workflow

This commit introduces a new automated workflow for deploying AWS Lambdas. The workflow deploys changes automatically as they are pushed to 'production' and 'lambda**' branches. Additionally, necessary dependencies for the deployment, like aws-cdk-lib and constructs, are also added to the lambda's requirements.

* Change deployment branch from 'production' to 'main'

The commit reflects changes made to the AWS Lambda and Analytics deployment yaml workflows. The workflows have been updated to initiate deployment upon pushing to the 'main' branch, instead of the earlier 'production' branch. This change aligns with the team's shift towards default branch naming conventions.

* Update .gitignore rules

The .gitignore file has been modified to ignore all files with the name "credentials" and the Python script "playground.py". This change is made to prevent accidental commit of sensitive credential files and temporary working scripts to the repository.

---------

Co-authored-by: prateek <prateek.patel@bricklane.com>