Modify stock strategy simulation to consider platform fees
This commit introduces an update to the stock strategy simulation. It now takes into account the platform fees in the trading strategy. Additionally, the code has been refactored for better readability, and a new 'weekly buy and hold' strategy has been added. The return of a portfolio can now also be tracked throughout the simulation.