Add detailed explanation of Double Poisson model
The explanation of the Double Poisson model used for game prediction has been added to the 'euro_2024_modelling.py' file. This includes a description of the model's key assumptions, limitations, and strengths, as well as how it applies to football match predictions. This update provides additional information to help users understand the logic behind the model's predictions.