Add project 'IMDB movie length analysis' to application
The project named 'IMDB movie length analysis' was added to the application. This involved creating a new Python script that fetches top grossing movies data each year from Boxoffice Mojo and movie details from IMDB. The main purpose of this project is to answer the question "Are movies getting longer?". This is done by comparing the mean runtimes of top grossing movies released between 1990-2000 to those released in 2022. The result is visualized and presented within the application by using the 'streamlit' and 'lets_plot' libraries.