Update euro_2024_modelling script with static values
The bayesian estimation in the euro_2024_modelling.py script is replaced with static values to speed up the UI. The iterative process that previously calculated these values was removed due to inaccuracies, and has been commented out for reference. The resulting values for "Attacks" and "Defences" are now provided directly to improve performance.

Signed-off-by: prateek <prteek@icloud.com>